+++ topiary-cli/src/main.rs
@@ -18,7 +18,7 @@ use rootcause::{
 };
 use tabled::{Table, settings::Style};
 use topiary_config::source::Source;
-use topiary_core::{ErrorSpan, Operation, SpanAttachment, check_query_coverage, formatter};
+use topiary_core::{Operation, SpanAttachment, check_query_coverage, formatter};
 
 use crate::{
     cli::Commands,
@@ -170,13 +170,7 @@ async fn run() -> CLIResult<()> {
                     output_format: format.into(),
                 },
             )
-            .map_err(|e| {
-                if let Some(filepath) = buf_input.get_ref().source().location().to_path() {
-                    return e.attach_filepath(filepath);
-                }
-                e
-            })
-            .context_to()?;
+            .attach_filepath(buf_input.get_ref().filepath())?;
         }
 
         Commands::Config {
@@ -244,14 +238,8 @@ async fn run() -> CLIResult<()> {
 
             let coverage_data =
                 check_query_coverage(&input_content, &language.query, &language.grammar)
-                    .attach_source(&input_content)
-                    .map_err(|e| {
-                        if let Some(filepath) = buf_input.get_ref().filepath() {
-                            return e.attach_filepath(filepath);
-                        }
-                        e
-                    })
-                    .context_to()?;
+                    .attach_source(input_content.as_str())
+                    .attach_filepath(buf_input.get_ref().filepath())?;
             let coverage_res = coverage_data.get_result();
 
             let query_source = NamedSource::new(