+++ topiary-core/src/lib.rs
@@ -221,10 +221,10 @@ pub enum Operation {
 ///   Ok(()) => {
 ///     let formatted = String::from_utf8(output).expect("valid utf-8");
 ///   }
-///   Err(FormatterError::Query(message, _)) => {
-///     panic!("Error in query file: {message}");
-///   }
-///   Err(_) => {
+///   Err(r) => {
+///     if let FormatterError::Query(message) = r.current_context() {
+///         panic!("Error in query file: {message}");
+///     }
 ///     panic!("An error occurred");
 ///   }
 /// }
@@ -415,7 +415,6 @@ mod tests {
             },
         );
 
-        dbg!(result.get_span().is_some());
         if let Some(range) = result
             .get_span()
             .and_then(|s| s.range)